TWINS-181: [classes] new autocreated class DOMAIN_USER to store user profile - fixes
